isf: simplify textures by using textureSize, now supported everywhere it matters
Also refactor utilities in RenderedISFNode a bit
